+++ modules/gis/test/psql/coerce_test.clj
@@ -28,3 +28,31 @@
 (deftest geojson->postgis->geojson
   (let [gj {:type :Point :coordinates [1.0 2.0]}]
     (is (= gj (coerce/postgis->geojson (coerce/geojson->postgis gj))))))
+
+(deftest geometry-collection-roundtrip
+  (let [gj {:type :GeometryCollection
+            :geometries [{:type :Point :coordinates [1.0 2.0]}
+                         {:type :LineString :coordinates [[0.0 0.0] [1.0 1.0]]}]}]
+    (testing "geojson->postgis builds a GeometryCollection"
+      (is (instance? net.postgis.jdbc.geometry.GeometryCollection
+                     (coerce/geojson->postgis gj))))
+    (testing "round-trips back to the same GeoJSON"
+      (is (= gj (coerce/postgis->geojson (coerce/geojson->postgis gj)))))))
+
+(deftest feature-and-feature-collection-conversion
+  (testing "a Feature converts to its underlying geometry"
+    (is (instance? net.postgis.jdbc.geometry.Point
+                   (coerce/geojson->postgis
+                    {:type :Feature
+                     :geometry {:type :Point :coordinates [1 2]}
+                     :properties {:name "x"}}))))
+  (testing "a FeatureCollection converts to a GeometryCollection of its geometries"
+    (let [gc (coerce/geojson->postgis
+              {:type :FeatureCollection
+               :features [{:type :Feature :geometry {:type :Point :coordinates [1 2]} :properties {}}
+                          {:type :Feature :geometry {:type :Point :coordinates [3 4]} :properties {}}]})]
+      (is (instance? net.postgis.jdbc.geometry.GeometryCollection gc))
+      (is (= {:type :GeometryCollection
+              :geometries [{:type :Point :coordinates [1.0 2.0]}
+                           {:type :Point :coordinates [3.0 4.0]}]}
+             (coerce/postgis->geojson gc))))))